project/requirements.txt
asyncio==3.4.3
attrs==20.3.0
autopep8==2.3.1
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
bcrypt==4.2.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.8.2
braceexpand==0.1.7
Brotli==1.1.0
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
check50==3.3.11
clang-format==19.1.2
cli50==8.0.1
click==8.1.7
colorama==0.4.6
compare50==1.2.6
contourpy==1.3.0
cryptography==43.0.3
cs50==9.4.0
CS50-VSIX-Client @ file:///opt/cs50/extensions/cs50vsix-client
cssselect2==0.7.0
cycler==0.12.1
djhtml==3.0.7
EditorConfig==0.12.4
Flask==3.0.3
Flask-Session==0.8.0
fonttools==4.54.1
greenlet==3.1.1
help50==3.0.5
html5lib==1.1
icdiff==2.0.7
idna==3.10
inflect==7.0.0
iniconfig==2.0.0
intervaltree==3.1.0
itsdangerous==2.2.0
jellyfish==0.11.2
Jinja2==3.1.4
jsbeautifier==1.15.1
kiwisolver==1.4.7
lib50==3.0.11
MarkupSafe==3.0.2
matplotlib==3.9.2
more-itertools==10.5.0
msgspec==0.18.6
mypy-extensions==1.0.0
natsort==8.4.0
numpy==1.26.4
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
psutil==6.1.0
ptyprocess==0.7.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==1.10.19
pydyf==0.11.0
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.1.0
pyphen==0.16.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
render50==9.2.7
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
style50==2.10.4
submit50==3.1.5
termcolor==1.1.0
tinycss2==1.4.0
tqdm==4.66.6
typeguard==4.4.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
weasyprint==62.3
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.6
wheel==0.44.0
zopfli==0.2.3.post1
